+++ docs/conf.py
@@ -26,16 +26,17 @@ release = ".".join(version.split(".")[:2])
 extensions = [
     "myst_parser",
     "sphinx.ext.autodoc",
+    "sphinx.ext.viewcode",
     "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
     "sphinx.ext.githubpages",
     "sphinx.ext.intersphinx",
-    "sphinx.ext.viewcode",
-    "nbsphinx",
     "sphinx_tippy",
     "sphinx_inline_tabs",
     "sphinx_copybutton",
     "versionwarning.extension",
+    "sphinx_autodoc_typehints",
+    "nbsphinx",
 ]
 source_suffix = {
     ".rst": "restructuredtext",
@@ -49,12 +50,17 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 html_static_path = ["_static"]
-html_theme = "furo"
+html_theme = "sphinx_book_theme"
 html_theme_options = {
-    "source_repository": "https://github.com/tanbro/emoji-data",
-    "source_branch": "main",
-    "source_directory": "docs/",
-    "top_of_page_button": "edit",
+    "path_to_docs": "docs/",
+    "repository_url": "https://github.com/tanbro/emoji-data",
+    "repository_branch": "master",
+    "use_download_button": True,
+    "use_fullscreen_button": True,
+    "use_repository_button": True,
+    "use_issues_button": True,
+    "use_download_button": True,
+    "show_toc_level": 2,
 }
 
 # -- Options for autodoc ----------------------------------------------------
@@ -64,7 +70,7 @@ html_theme_options = {
 
 # Automatically extract typehints when specified and place them in
 # descriptions of the relevant function/method.
-autodoc_typehints = "description"
+autodoc_typehints = "both"
 
 # Don't show class signature with the class' name.
 # autodoc_class_signature = "separated"